refactor(components): New components classes
New RowComponent and IterableComponent class

refactor(components): New components classes

New RowComponent and IterableComponent class
